Refactor tests and isolate bug
