+++ core/src/test/scala/ai/lum/odinson/state/TestMockState.scala
@@ -7,9 +7,8 @@ class TestMockState extends OdinsonTest {
   val docGummy = getDocument("becky-gummy-bears-v2")
 
   val eeGummy = extractorEngineWithSpecificState(docGummy, "mock")
-  val eeGummyMemory = extractorEngineWithSpecificState(docGummy, "memory")
 
-  "MockState" should "return mentions" in {
+  "MockState" should "return mentions that don't rely on state-based retrieval" in {
     val rules = """
           |rules:
           |  - name: gummy-rule
@@ -44,44 +43,6 @@ class TestMockState extends OdinsonTest {
     mentions should have size (4)
 
     getMentionsWithLabel(mentions, "GummyBear") should have size (0)
-
-  }
-
-  "MemoryState" should "return mentions" in {
-    val rules = """
-        |rules:
-        |  - name: gummy-rule
-        |    label: Bear
-        |    type: basic
-        |    priority: 1
-        |    pattern: |
-        |      gummy
-        |
-        |  - name: eating-rule
-        |    label: Consumption
-        |    type: event
-        |    priority: 2
-        |    pattern: |
-        |      trigger = [lemma=eat]
-        |      subject: ^NP = >nsubj []
-        |      object: ^NP = >dobj []
-        |
-        |  - name: nomatch-rule
-        |    label: Gummy
-        |    type: event
-        |    priority: 2
-        |    pattern: |
-        |      trigger = bears
-        |      arg: Bear = >amod
-       """.stripMargin
-
-    val extractors = eeGummyMemory.ruleReader.compileRuleString(rules)
-    val mentions = eeGummyMemory.extractMentions(extractors).toArray
-
-    // the 3 main extractions + 2 promoted args
-    mentions should have size (5)
-
-    getMentionsWithLabel(mentions, "Gummy") should have size (1)
 
   }
